api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi[standard]>=0.135.0",
    "langchain>=1.2.10",
    "langgraph-checkpoint-postgres>=3.0.4",
    "psycopg[binary]>=3.3.3",
    "pydantic-settings>=2.13.1",
]
